chore: upgrade dependencies to latest major versions
Motivation:
- To keep the project up to date with the latest features, performance improvements, and security fixes from dependencies.

Changes:
- Upgraded all JavaScript dependencies to their latest versions in cat-launcher/package.json.
- Upgraded Rust dependencies in cat-launcher/src-tauri/Cargo.toml and cat-launcher/cat-macros/Cargo.toml to their latest major versions.
- Fixed breaking changes in the backend:
  - Updated rusqlite usage to explicitly handle u64 and usize casting to i64 for SQLite compatibility.
  - Updated posthog-rs initialization to use the new async Client API and the renamed host() method in ClientOptionsBuilder.
- Kept reqwest at 0.12.25 to maintain compatibility with the downloader crate, which is a git dependency.

+++ cat-launcher/src-tauri/src/variants/repository/sqlite_game_variant_order_repository.rs
@@ -82,7 +82,7 @@ impl GameVariantOrderRepository for SqliteGameVariantOrderRepository {
             for (i, variant) in variants.iter().enumerate() {
                 tx.execute(
                     "INSERT INTO game_variant_order (game_variant, sort_order) VALUES (?1, ?2)",
-                    rusqlite::params![variant.to_string(), i],
+                    rusqlite::params![variant.to_string(), i as i64],
                 )
                 .map_err(|e| UpdateGameVariantOrderError::Update(Box::new(e)))?;
             }
